+++ src/App.js
@@ -1,30 +1,13 @@
 import React from 'react';
-const todoList = [
-  {
-    id:1,
-    title: "Complete assigment"
-  },
-  {
-    id:2,
-    title: "Run assigment"
-  },
-  {
-    id:3,
-    title: "Check assigment"
-  }
-];
+import TodoList from './TodoList';
+import AddTodoForm from './AddTodoForm';
 
 function App() {
   return (
     <div>
       <h1>Todo List</h1>
-      <ul>
-        {todoList.map(function(item) {
-          return <li key={item.id}> 
-            <span>{item.title}</span> 
-          </li>
-        })}
-      </ul>
+      <AddTodoForm />
+      <TodoList />
     </div>
   );
 }
